// src/light-app/store/types.ts
export enum LightAppActionType {
  CreateApp = 'createApp',
  CreateForm = 'createForm',
  UpdateForm = 'updateForm',
  AsyncUpdateFormName = 'asyncUpdateFormName',
  RenameSheet = 'renameSheet',
  GetFormControls = 'getFormControls',
  GetFormDataLists = 'getFormDataLists'
}
export enum LightAppMutationType {
  AppCode = 'appCode',
  FormCode = 'formCode',
  HandleRenameSheet = 'handleRenameSheet',
  CreateForm = 'createForm',
  UpdateForm = 'updateForm',
  UpdateFormName = 'updateFormName',
  UpdateFormControls = 'updateFormControls',
  FormControlsDele = 'formControlsDele',
  MoveFormControls = 'moveFormControls',
  FormControls = 'formControls',
  SelectedControl = 'selectedControl',
  SelectFormObjectId = 'selectFormObjectId',
  UpdateControlCodeList = 'UpdateControlCodeList',
  IsWorkflow = 'isWorkflow',
  HandleUniqueControls = 'handleUniqueControls',
  UpdateColsNameList = 'updateColsNameList',
  HandleFormDataLists = 'handleFormDataLists'
}